# apps/wanaku-barn-backend/src/main/webui/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
      "AuditEvent" : {
        "type" : "object",
        "properties" : {
          "schema_version" : {
            "type" : "string"
          },
          "event_id" : {
            "type" : "string"
          },
          "stream_id" : {
            "type" : "string"
          },
          "sequence" : {
            "type" : "integer",
            "format" : "int64"
          },
          "category" : {
            "type" : "string"
          },
          "decision" : {
            "type" : "string"
          },
          "correlation_id" : {
            "type" : "string"
          },
          "request_id" : {
            "type" : "string"
          },
          "actor" : {
            "type" : "string"
          },
          "protocol" : {
            "type" : "string"
          },
          "operation" : {
            "type" : "string"
          },
          "target_type" : {
            "type" : "string"
          },
          "target" : {
            "type" : "string"
          },
          "reason_code" : {
            "type" : "string"
          },
          "explanation" : {
            "type" : "string"
          },
          "policy_revision" : {
            "type" : "string"
          },
          "response_status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "duration_ms" : {
            "type" : "integer",
            "format" : "int64"
          },
          "redaction" : {
            "$ref" : "#/components/schemas/RedactionMetadata"
          },
          "attributes" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "coverage_complete" : {
            "type" : "boolean"
          },
          "dropped_events" : {
            "type" : "integer",
            "format" : "int64"
          },
          "timestamp" : {
            "type" : "string"
          }
        }
      },
...
      "AuditPage" : {
        "type" : "object",
        "properties" : {
          "events" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AuditEvent"
            }
          },
          "offset" : {
            "type" : "integer",
            "format" : "int32"
          },
          "limit" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "RedactionMetadata" : {
        "type" : "object",
        "properties" : {
          "redacted_fields" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "payload_captured" : {
            "type" : "boolean"
          },
          "payload_truncated" : {
            "type" : "boolean"
          }
        }
      },
...
      "WanakuError" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "WanakuResponseAuditEvent" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "$ref" : "#/components/schemas/AuditEvent"
          }
        }
      },
...
      "WanakuResponseAuditPage" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "$ref" : "#/components/schemas/WanakuError"
          },
          "data" : {
            "$ref" : "#/components/schemas/AuditPage"
          }
        }
      },
...
    }
  },
  "paths" : {
    "/api/v1/audit/events" : {
      "get" : {
        "parameters" : [ {
          "name" : "actor",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "decision",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 100
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "operation",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "reason_code",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "target",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "target_type",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseAuditPage"
                }
              }
            }
          }
        },
        "summary" : "List",
        "tags" : [ "Audit Resource" ]
      }
    },
    "/api/v1/audit/events/{id}" : {
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WanakuResponseAuditEvent"
                }
              }
            }
          }
        },
        "summary" : "Get",
        "tags" : [ "Audit Resource" ]
      }
    },
...
  },
...
  "info" : {
    "title" : "wanaku-barn-backend API",
    "version" : "0.3.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8180",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8180",
    "description" : "Auto generated value"
  } ]
}